Use doubles instead of floats

+++ Point_set_processing_3/test/Point_set_processing_3/wlop_simplify_and_regularize_test.cpp
@@ -28,7 +28,7 @@
 // ----------------------------------------------------------------------------
 
 // kernel
-typedef CGAL::Simple_cartesian<float> Kernel;
+typedef CGAL::Simple_cartesian<double> Kernel;
 
 // Simple geometric types
 typedef Kernel::FT FT;
